test(#1491): pin F1/F2/F3/F5 from both sides; split the comparator's cases out
Adds the four review findings as unit cases, each pinned from both sides:
the CSV empty-member rule (top level vs inside a container), a text map key
compared by JSON kind while an int-keyed map still pairs across spellings, the
format-scoped UDT representation, and a field-scoped skip that keeps comparing
its sibling fields plus the stale-skip report.

golden_value_compare.rs had grown to 1690 lines, so its `mod tests` moves to
golden_value_compare_tests.rs (campsite rule, #1135): 901 + 800.
