+++ cqlite-cli/tests/support/golden_value_compare.rs
@@ -897,497 +897,5 @@ pub fn stage_single_table(dest: &Path, keyspace: &str, fixture: &Path) -> Result
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use serde_json::json;
-
-    fn row(pairs: &[(&str, Value)]) -> Row {
-        pairs
-            .iter()
-            .map(|(k, v)| ((*k).to_string(), v.clone()))
-            .collect()
-    }
-
-    /// A schema for the unit cases, parsed from DDL exactly as the lane parses the
-    /// committed files — so these cases exercise the real authority, not a mock.
-    fn schema_of(ddl: &str, table: &str) -> TableSchema {
-        match super::super::schema::from_ddl(ddl, table) {
-            Ok(schema) => schema,
-            Err(why) => panic!("{table}: {why}"),
-        }
-    }
-
-    fn set_schema() -> TableSchema {
-        schema_of("CREATE TABLE t (id int PRIMARY KEY, s set<text>);", "t")
-    }
-
-    /// The refusal PATH, not just the predicate: no corpus fixture carries a
-    /// `, `-bearing collection member, so without this the wiring from
-    /// `csv_container::ambiguity` to the census counters never executes and the
-    /// lane's `0 REFUSED` line would be unfalsifiable.
-    #[test]
-    fn a_csv_unrepresentable_container_is_refused_and_named() {
-        let schema = set_schema();
-        let golden = vec![row(&[("id", json!(1)), ("s", json!(["a, b"]))])];
-        // The CLI text is IRRELEVANT to the refusal: it is decided from the
-        // golden alone, so the defect under test can never cause it.
-        let cli = vec![row(&[("id", json!("1")), ("s", json!("{a, b}"))])];
-        let report = compare_rows(&golden, &cli, &schema, &["id"], &[], &[], Egress::Csv);
-
-        assert!(
-            report.diffs.is_empty(),
-            "unexpected diffs: {:?}",
-            report.diffs
-        );
-        assert_eq!(report.ambiguous_container_cells, 1);
-        assert_eq!(
-            report.container_cells, 0,
-            "a refused cell is not a compared one"
-        );
-        assert_eq!(report.compared_cells, 1, "`id` is still compared");
-        assert_eq!(report.ambiguity_reasons.len(), 1);
-        assert!(
-            report.ambiguity_reasons[0].starts_with("s ("),
-            "the refusal must name its column: {:?}",
-            report.ambiguity_reasons
-        );
-    }
-
-    /// A representable container is compared, and a wrong member fails. Pins the
-    /// other side of the same branch so "refused" can never quietly become the
-    /// default.
-    #[test]
-    fn a_representable_container_is_compared_and_a_wrong_member_fails() {
-        let schema = set_schema();
-        let golden = vec![row(&[("id", json!(1)), ("s", json!(["a", "b"]))])];
-        let good = vec![row(&[("id", json!("1")), ("s", json!("{a, b}"))])];
-        let report = compare_rows(&golden, &good, &schema, &["id"], &[], &[], Egress::Csv);
-        assert!(
-            report.diffs.is_empty(),
-            "unexpected diffs: {:?}",
-            report.diffs
-        );
-        assert_eq!(report.container_cells, 1);
-        assert_eq!(report.ambiguous_container_cells, 0);
-
-        let bad = vec![row(&[("id", json!("1")), ("s", json!("{a, c}"))])];
-        let report = compare_rows(&golden, &bad, &schema, &["id"], &[], &[], Egress::Csv);
-        assert_eq!(
-            report.diffs.len(),
-            1,
-            "a wrong member must fail: {:?}",
-            report.diffs
-        );
-        assert!(report.diffs[0].contains(".s:"), "{:?}", report.diffs);
-    }
-
-    /// A CLI cell that is not the grammar at all is a DIVERGENCE, not a refusal.
-    #[test]
-    fn an_unparseable_container_is_reported_not_refused() {
-        let schema = set_schema();
-        let golden = vec![row(&[("id", json!(1)), ("s", json!(["a", "b"]))])];
-        let cli = vec![row(&[("id", json!("1")), ("s", json!("a, b"))])];
-        let report = compare_rows(&golden, &cli, &schema, &["id"], &[], &[], Egress::Csv);
-        assert_eq!(report.ambiguous_container_cells, 0);
-        assert_eq!(report.diffs.len(), 1, "{:?}", report.diffs);
-        assert!(
-            report.diffs[0].contains("unparseable CSV container"),
-            "{:?}",
-            report.diffs
-        );
-    }
-
-    // =======================================================================
-    // The column set is the DDL's, and only the golden may default to null
-    // =======================================================================
-
-    const ABSENT_DDL: &str =
-        "CREATE TABLE t (pk int, ck int, anchor text, reg text, PRIMARY KEY (pk, ck));";
-
-    fn absent_schema() -> TableSchema {
-        schema_of(ABSENT_DDL, "t")
-    }
-
-    fn absent_golden() -> Vec<Row> {
-        // The `sstabledump` shape of `test_types.nb_absent_vs_null_regular` row 1:
-        // `reg` was never written, so the physical dump simply has no such cell.
-        vec![row(&[
-            ("pk", json!(1)),
-            ("ck", json!(1)),
-            ("anchor", json!("anchor_absent")),
-        ])]
-    }
-
-    /// The property `nb_absent_vs_null_regular` exists for: an absent cell must be
-    /// RENDERED as null. Reading an omitted egress column as null (the first cut of
-    /// this file) made it unfalsifiable.
-    #[test]
-    fn a_column_the_egress_omits_is_a_named_failure() {
-        let schema = absent_schema();
-        let golden = absent_golden();
-
-        let rendered = vec![row(&[
-            ("pk", json!(1)),
-            ("ck", json!(1)),
-            ("anchor", json!("anchor_absent")),
-            ("reg", Value::Null),
-        ])];
-        let report = compare_rows(
-            &golden,
-            &rendered,
-            &schema,
-            &["pk"],
-            &["ck"],
-            &[],
-            Egress::Json,
-        );
-        assert!(
-            report.diffs.is_empty(),
-            "an absent golden cell rendered as null is the expected outcome: {:?}",
-            report.diffs
-        );
-        assert_eq!(
-            report.compared_cells, 4,
-            "every declared column is compared"
-        );
-
-        // The regression: the egress drops the column entirely.
-        let omitted = vec![row(&[
-            ("pk", json!(1)),
-            ("ck", json!(1)),
-            ("anchor", json!("anchor_absent")),
-        ])];
-        let report = compare_rows(
-            &golden,
-            &omitted,
-            &schema,
-            &["pk"],
-            &["ck"],
-            &[],
-            Egress::Json,
-        );
-        assert_eq!(report.diffs.len(), 1, "{:?}", report.diffs);
-        assert!(
-            report.diffs[0].contains("reg") && report.diffs[0].contains("absent from the"),
-            "the failure must name the omitted column: {:?}",
-            report.diffs
-        );
-    }
-
-    /// The mirror image: a column the DDL does not declare must not pass just
-    /// because it holds `null`.
-    #[test]
-    fn a_spurious_extra_column_is_a_named_failure() {
-        let schema = absent_schema();
-        let golden = absent_golden();
-        let extra = vec![row(&[
-            ("pk", json!(1)),
-            ("ck", json!(1)),
-            ("anchor", json!("anchor_absent")),
-            ("reg", Value::Null),
-            ("ghost", Value::Null),
-        ])];
-        let report = compare_rows(
-            &golden,
-            &extra,
-            &schema,
-            &["pk"],
-            &["ck"],
-            &[],
-            Egress::Json,
-        );
-        assert_eq!(report.diffs.len(), 1, "{:?}", report.diffs);
-        assert!(
-            report.diffs[0].contains("ghost") && report.diffs[0].contains("does not declare"),
-            "the failure must name the undeclared column: {:?}",
-            report.diffs
-        );
-    }
-
-    /// A value where the golden has no cell at all is still a divergence — the
-    /// golden's absence is an expected NULL, not a wildcard.
-    #[test]
-    fn a_value_where_the_golden_has_no_cell_still_fails() {
-        let schema = absent_schema();
-        let golden = absent_golden();
-        let invented = vec![row(&[
-            ("pk", json!(1)),
-            ("ck", json!(1)),
-            ("anchor", json!("anchor_absent")),
-            ("reg", json!("invented")),
-        ])];
-        let report = compare_rows(
-            &golden,
-            &invented,
-            &schema,
-            &["pk"],
-            &["ck"],
-            &[],
-            Egress::Json,
-        );
-        assert_eq!(report.diffs.len(), 1, "{:?}", report.diffs);
-        assert!(
-            report.diffs[0].contains(".reg:") && report.diffs[0].contains("invented"),
-            "{:?}",
-            report.diffs
-        );
-    }
-
-    /// A golden cell for a column the named schema does not declare means the
-    /// expectation itself is stale, so it is a failure rather than silent coverage.
-    #[test]
-    fn a_golden_cell_for_an_undeclared_column_is_a_named_failure() {
-        let schema = absent_schema();
-        let golden = vec![row(&[
-            ("pk", json!(1)),
-            ("ck", json!(1)),
-            ("anchor", json!("a")),
-            ("dropped", json!("x")),
-        ])];
-        let cli = vec![row(&[
-            ("pk", json!(1)),
-            ("ck", json!(1)),
-            ("anchor", json!("a")),
-            ("reg", Value::Null),
-        ])];
-        let report = compare_rows(&golden, &cli, &schema, &["pk"], &["ck"], &[], Egress::Json);
-        assert!(
-            report
-                .diffs
-                .iter()
-                .any(|d| d.contains("dropped") && d.contains("does not declare")),
-            "{:?}",
-            report.diffs
-        );
-    }
-
-    /// A shape divergence is reported ONCE per column, not once per row — the
-    /// tables in this lane run to 900 rows.
-    #[test]
-    fn a_column_shape_divergence_is_reported_once_per_column() {
-        let schema = absent_schema();
-        let golden: Vec<Row> = (1..=5)
-            .map(|i| {
-                row(&[
-                    ("pk", json!(1)),
-                    ("ck", json!(i)),
-                    ("anchor", json!("a")),
-                    ("reg", json!("v")),
-                ])
-            })
-            .collect();
-        let cli: Vec<Row> = (1..=5)
-            .map(|i| row(&[("pk", json!(1)), ("ck", json!(i)), ("anchor", json!("a"))]))
-            .collect();
-        let report = compare_rows(&golden, &cli, &schema, &["pk"], &["ck"], &[], Egress::Json);
-        assert_eq!(
-            report.diffs.len(),
-            1,
-            "5 rows missing the same column must report once: {:?}",
-            report.diffs
-        );
-    }
-
-    /// A declared `skip_columns` entry still suppresses its column, so the
-    /// measured-divergence gaps keep working.
-    #[test]
-    fn a_declared_skip_column_is_not_required_to_be_rendered() {
-        let schema = absent_schema();
-        let golden = absent_golden();
-        let omitted = vec![row(&[
-            ("pk", json!(1)),
-            ("ck", json!(1)),
-            ("anchor", json!("anchor_absent")),
-        ])];
-        let report = compare_rows(
-            &golden,
-            &omitted,
-            &schema,
-            &["pk"],
-            &["ck"],
-            &["reg"],
-            Egress::Json,
-        );
-        assert!(
-            report.diffs.is_empty(),
-            "a declared skip must stay declared: {:?}",
-            report.diffs
-        );
-        assert_eq!(report.compared_cells, 3);
-    }
-
-    // =======================================================================
-    // Types come from the DDL: a numeric-looking text is NOT a number
-    // =======================================================================
-
-    // One DDL per shape: the comparison now REQUIRES every declared column to be
-    // rendered, so a schema carrying columns a case does not exercise would fail
-    // for the right reason in the wrong test.
-    const TEXT_DDL: &str = "CREATE TABLE t (id int PRIMARY KEY, zip text);";
-    const NUM_DDL: &str = "CREATE TABLE t (id int PRIMARY KEY, n int);";
-    const UDT_MAP_DDL: &str = "CREATE TYPE address (street text, city text, zip text); \
-         CREATE TABLE t (id int PRIMARY KEY, ma map<text, frozen<address>>);";
-    const INT_MAP_DDL: &str = "CREATE TABLE t (id int PRIMARY KEY, mi map<int, text>);";
-
-    /// BLOCKER 2: a CQL `text` value holding `\"22201\"` must NOT compare equal to
-    /// the JSON number `22201`, and `\"00000\"` must not equal `\"0\"`.
-    #[test]
-    fn a_numeric_looking_text_column_is_compared_exactly() {
-        let schema = schema_of(TEXT_DDL, "t");
-        let golden = vec![row(&[("id", json!(1)), ("zip", json!("22201"))])];
-
-        let same = vec![row(&[("id", json!(1)), ("zip", json!("22201"))])];
-        let report = compare_rows(&golden, &same, &schema, &["id"], &[], &[], Egress::Json);
-        assert!(report.diffs.is_empty(), "{:?}", report.diffs);
-
-        for wrong in [json!(22201), json!("22201.0"), json!("022201")] {
-            let cli = vec![row(&[("id", json!(1)), ("zip", wrong.clone())])];
-            let report = compare_rows(&golden, &cli, &schema, &["id"], &[], &[], Egress::Json);
-            assert_eq!(
-                report.diffs.len(),
-                1,
-                "text {wrong} must not equal the golden text \"22201\": {:?}",
-                report.diffs
-            );
-            assert!(report.diffs[0].contains(".zip:"), "{:?}", report.diffs);
-        }
-
-        // Zero padding, the second half of the finding.
-        let padded = vec![row(&[("id", json!(1)), ("zip", json!("00000"))])];
-        let stripped = vec![row(&[("id", json!(1)), ("zip", json!("0"))])];
-        let report = compare_rows(&padded, &stripped, &schema, &["id"], &[], &[], Egress::Json);
-        assert_eq!(
-            report.diffs.len(),
-            1,
-            "\"00000\" must not equal \"0\": {:?}",
-            report.diffs
-        );
-    }
-
-    /// The same strictness one level in: a `text` UDT field nested inside a map
-    /// value. This is the shape the `udt_collections` fixture actually carries
-    /// (`ma frozen<map<text, frozen<address>>>`, zip `\"22201\"`/`\"00000\"`).
-    #[test]
-    fn a_numeric_looking_text_udt_field_is_compared_exactly() {
-        let schema = schema_of(UDT_MAP_DDL, "t");
-        let golden = vec![row(&[
-            ("id", json!(1)),
-            (
-                "ma",
-                json!({"home": {"street": "1 Navy Way", "city": "Arlington", "zip": "00000"}}),
-            ),
-        ])];
-        let cli_ok = vec![row(&[
-            ("id", json!(1)),
-            (
-                "ma",
-                json!([{"key": "home", "value": {"_type": "address", "street": "1 Navy Way",
-                        "city": "Arlington", "zip": "00000"}}]),
-            ),
-        ])];
-        let report = compare_rows(&golden, &cli_ok, &schema, &["id"], &[], &[], Egress::Json);
-        assert!(report.diffs.is_empty(), "{:?}", report.diffs);
-        assert_eq!(report.container_cells, 1);
-
-        for wrong in [json!(0), json!("0")] {
-            let cli = vec![row(&[
-                ("id", json!(1)),
-                (
-                    "ma",
-                    json!([{"key": "home", "value": {"_type": "address",
-                            "street": "1 Navy Way", "city": "Arlington", "zip": wrong}}]),
-                ),
-            ])];
-            let report = compare_rows(&golden, &cli, &schema, &["id"], &[], &[], Egress::Json);
-            assert_eq!(
-                report.diffs.len(),
-                1,
-                "a nested text zip {wrong} must not equal \"00000\": {:?}",
-                report.diffs
-            );
-            assert!(report.diffs[0].contains("zip"), "{:?}", report.diffs);
-        }
-    }
-
-    /// The normalization that must SURVIVE: a numeric column's golden spelling is a
-    /// string (a partition key, a collection path) and the CLI's is a number.
-    #[test]
-    fn a_numeric_column_still_compares_across_spellings() {
-        let schema = schema_of(NUM_DDL, "t");
-        let golden = vec![row(&[("id", json!("1")), ("n", json!("-5"))])];
-        let cli = vec![row(&[("id", json!(1)), ("n", json!(-5))])];
-        let report = compare_rows(&golden, &cli, &schema, &["id"], &[], &[], Egress::Json);
-        assert!(
-            report.diffs.is_empty(),
-            "a numeric column must still pair a string spelling with a number: {:?}",
-            report.diffs
-        );
-
-        // And a WRONG number still fails.
-        let wrong = vec![row(&[("id", json!(1)), ("n", json!(-6))])];
-        let report = compare_rows(&golden, &wrong, &schema, &["id"], &[], &[], Egress::Json);
-        assert_eq!(report.diffs.len(), 1, "{:?}", report.diffs);
-    }
-
-    /// Map KEYS are canonicalized under the declared KEY type: numeric for
-    /// `map<int,…>` (the dump renders every path as a string), exact for
-    /// `map<text,…>`.
-    #[test]
-    fn map_keys_are_canonicalized_under_the_declared_key_type() {
-        let schema = schema_of(INT_MAP_DDL, "t");
-        let golden = vec![row(&[("id", json!(1)), ("mi", json!({"-5": "v"}))])];
-        let cli = vec![row(&[
-            ("id", json!(1)),
-            ("mi", json!([{"key": -5, "value": "v"}])),
-        ])];
-        let report = compare_rows(&golden, &cli, &schema, &["id"], &[], &[], Egress::Json);
-        assert!(report.diffs.is_empty(), "{:?}", report.diffs);
-
-        let wrong = vec![row(&[
-            ("id", json!(1)),
-            ("mi", json!([{"key": -6, "value": "v"}])),
-        ])];
-        let report = compare_rows(&golden, &wrong, &schema, &["id"], &[], &[], Egress::Json);
-        assert_eq!(report.diffs.len(), 1, "{:?}", report.diffs);
-
-        // A text-keyed map compares its keys exactly.
-        let schema = schema_of(UDT_MAP_DDL, "t");
-        let golden = vec![row(&[
-            ("id", json!(1)),
-            (
-                "ma",
-                json!({"00000": {"street": "s", "city": "c", "zip": "z"}}),
-            ),
-        ])];
-        let wrong = vec![row(&[
-            ("id", json!(1)),
-            (
-                "ma",
-                json!([{"key": 0, "value": {"street": "s", "city": "c", "zip": "z"}}]),
-            ),
-        ])];
-        let report = compare_rows(&golden, &wrong, &schema, &["id"], &[], &[], Egress::Json);
-        assert_eq!(
-            report.diffs.len(),
-            1,
-            "a text map key \"00000\" must not equal the number 0: {:?}",
-            report.diffs
-        );
-    }
-
-    /// CSV keeps the same type rule: a `text` cell is exact even though every CSV
-    /// field arrives as text.
-    #[test]
-    fn the_csv_lane_uses_the_declared_types_too() {
-        let schema = schema_of(TEXT_DDL, "t");
-        let golden = vec![row(&[("id", json!(1)), ("zip", json!("00000"))])];
-        let ok = vec![row(&[("id", json!("1")), ("zip", json!("00000"))])];
-        let report = compare_rows(&golden, &ok, &schema, &["id"], &[], &[], Egress::Csv);
-        assert!(report.diffs.is_empty(), "{:?}", report.diffs);
-
-        let wrong = vec![row(&[("id", json!("1")), ("zip", json!("0"))])];
-        let report = compare_rows(&golden, &wrong, &schema, &["id"], &[], &[], Egress::Csv);
-        assert_eq!(report.diffs.len(), 1, "{:?}", report.diffs);
-        assert!(report.diffs[0].contains(".zip:"), "{:?}", report.diffs);
-    }
-}
+#[path = "golden_value_compare_tests.rs"]
+mod tests;
